refactor: split public category tree views (#352)
